+++ package.json
@@ -26,7 +26,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "sass": "sass src/:src/ --watch --no-source-map",
-    "server": "json-server --watch db.json --port 5000"
+    "server": "json-server --watch db.json --port 5000 --delay 2000"
   },
   "eslintConfig": {
     "extends": [
